python_service/requirements.txt
flask==3.0.0
qiskit==0.45.0
qiskit-optimization==0.6.0
mysql-connector-python==8.2.0

